# api.yaml
openapi: 3.0.3
info:
  title: cli-mgt-svc
  version: "2.4"
  description: |
    REST API for user management, Network Element (NE) management,
    NE connection config, NETCONF config backup, command history, and bulk import.

    **Permission model:** Users have a single permission derived from `account_type`:
    `0` (SuperAdmin) and `1` (Admin) → `admin`; `2` (Normal) → `user`.
    Only `admin` users may call management APIs.

    **Authentication:** Most endpoints require `Authorization: Basic <jwt_token>`.
    Token obtained from `POST /aa/authenticate` — the `response_data` field already includes the `Basic ` prefix, paste it as-is.
servers:
  - url: http://localhost:3000
    description: Local
security: []
...
paths:
...
  /aa/heath-check-db:
    get:
      tags: [System]
      summary: Health check (DB ping) — /aa alias
      description: Same as `GET /health` but mounted under the `/aa` prefix for compatibility with internal callers.
      operationId: heathCheckDb
      security: []
      responses:
        "200":
          description: Service healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Authentication ───────────────────────────────────────────────────────────
  /aa/authenticate:
    post:
      tags: [Authentication]
      summary: Login — get JWT
      operationId: authenticate
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              username: anhdt195
              password: "123"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /aa/validate-token:
    post:
      tags: [Authentication]
      summary: Validate token
      operationId: validateToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateTokenRequest"
      responses:
        "200":
          description: Token valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidateTokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /aa/change-password/:
    post:
      tags: [Authentication]
      summary: Change own password (requires old password)
      description: |
        Any authenticated user (including Normal role) may call this endpoint to
        change their own password. The `username` in the body must match the
        authenticated user — otherwise the request returns 401.
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Wrong old password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /aa/import/:
    post:
      tags: [Import]
      summary: Bulk import users, NEs, roles, mappings and NE configs from plain text
      operationId: bulkImport
      description: |
        Body is plain text in section format:

        ```
        [users]
        username,password,email
        admin,admin123,admin@vht.com

        [nes]
        ne_name,site_name,namespace,command_url,conf_mode,conf_master_ip,conf_port_master_ssh,conf_username,conf_password,description
        HTSMF01,HCM,hcm-5gc,http://10.10.1.1:8080,SSH,10.10.1.1,22,admin,admin,HCM SMF Node 01

        [user_roles]
        username,permission
        operator1,admin
        viewer1,user

        [user_nes]
        username,ne_name
        operator1,HTSMF01
        ```

        Existing users are skipped (idempotent). `[user_roles]` sets account_type (admin=1, user=2).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
      responses:
        "200":
          description: Import results (one entry per row)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Admin (Frontend API) ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Paste the `response_data` value from `POST /aa/authenticate` directly.
        The token already contains the `Basic ` prefix — paste it as-is.
        Example: `Basic eyJhbGciOiJIUzI1NiJ9...`

  # ── Reusable responses ───────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: Invalid request body / params
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"
...
    InternalError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/MessageResponse"

  # ── Schemas ──────────────────────────────────────────────────────────────────
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        response_data:
          type: string
          description: "Token string including 'Basic ' prefix (e.g. 'Basic eyJ...'). Use this value directly as the Authorization header."
          example: "Basic eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        response_code:
          type: string
          example: "200"
        system_type:
          type: string
          example: 5GC
    ValidateTokenRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: "The value of response_data from /aa/authenticate (already includes 'Basic ' prefix)"
    ValidateTokenResponse:
      type: object
      properties:
        username:
          type: string
        roles:
          type: string
          description: Permission derived from account_type — "admin" (type 0/1) or "user" (type 2)
    ChangePasswordRequest:
      type: object
      required: [username, old_password, new_password]
      properties:
        username:
          type: string
        old_password:
          type: string
        new_password:
          type: string
...
    ImportResult:
      type: object
      properties:
        type:
          type: string
          description: "Section type: user, ne, role, user_role, user_ne, ne_config"
          example: user
        name:
          type: string
          description: Identifier of the imported row
          example: admin
        status:
          type: string
          enum: [ok, skip, error]
        detail:
          type: string
          description: Additional info (reason for skip/error)

    # ── Subscribers ───────────────────────────────────────────────────────────
...